+++ toolchain/scripts/modules/watch/folder.ps1
@@ -3,7 +3,7 @@ function Watch-Folder {
    [Parameter(Mandatory=$true)][String] $Path
   )
   if ((Test-Path -Path $Path) -eq $false) { Write-Host "error - $Path not found" -ForegroundColor Red; exit 1 }
-  Logger-Step -Message "waiting change in $Path"
+  Write-Host "Waiting for changes in $Path" -ForegroundColor Yellow
 
   $sizeLast = (Get-ChildItem -Path $Path | Measure-Object -Sum Length | Select-Object Count, Sum).Sum
   $sizeCurrent = $sizeLast