+++ toolchain/scripts/modules/mak/run/mame.ps1
@@ -1,0 +1,7 @@
+function MakRunMame {
+  Build-Sprite
+  Build-Program
+  Build-ISO
+  Build-Mame
+  Start-Mame
+}