+++ toolchain/scripts/modules/write/iso.ps1
@@ -11,14 +11,17 @@ function Write-ISO {
     [Parameter(Mandatory=$true)][String] $PathCDTemplate
   )
 
-  Logger-Step -Message "compiling ISO"
+  Write-Host "Compiling ISO $OutputFile" -ForegroundColor Yellow
 
   & mkisofs.exe -o $OutputFile -pad $PathISOBuildFolder
 
   if ((Test-Path -Path $OutputFile) -eq $true) {
-    Logger-Success -Message "builded ISO $OutputFile"
+    Write-Host "Builded ISO $OutputFile" -ForegroundColor Green
     Write-Host ""
-  } else { Logger-Error -Message "$OutputFile was not generated" }
+  } else {
+    Write-Host "$OutputFile was not generated" -ForegroundColor Red
+    exit 1
+  }
 }
 
 function Write-Cache {
@@ -33,18 +36,27 @@ function Write-Cache {
     if ($Manifest.manifest.dependencies.cdTemplate.url) {
       Install-Component `
         -URL $Manifest.manifest.dependencies.cdTemplate.url `
-        -PathDownload $buildConfig.pathSpool `
+        -PathDownload "$($Config.project.buildPath)\spool" `
         -PathInstall $Manifest.manifest.dependencies.cdTemplate.path
     } else {
-      Install-Component -URL "$($buildConfig.baseURL)/neobuild-cd_template.zip" -PathDownload $buildConfig.pathSpool -PathInstall $buildConfig.pathNeocore
+      Install-Component -URL "$BaseURL/neobuild-cd_template.zip" -PathDownload "$($Config.project.buildPath)\spool" -PathInstall $Config.project.buildPath
     }
   }
 
   if (Test-Path -Path $PathISOBuildFolder) { Remove-Item $PathISOBuildFolder -Recurse -Force }
   if (-Not(Test-Path -Path $PathISOBuildFolder)) { mkdir -Path $PathISOBuildFolder | Out-Null }
-  if (-Not(Test-Path -Path $PathCDTemplate)) { Logger-Error -Message "$PathCDTemplate not found" }
-  if (-Not(Test-Path -Path $PRGFile)) { Logger-Error -Message "$PRGFile not found" }
-  if (-Not(Test-Path -Path $SpriteFile)) { Logger-Error -Message "$SpriteFile not found" }
+  if (-Not(Test-Path -Path $PathCDTemplate)) {
+    Write-Host "$PathCDTemplate not found" -ForegroundColor Red
+    exit 1
+  }
+  if (-Not(Test-Path -Path $PRGFile)) {
+    Write-Host "$PRGFile not found" -ForegroundColor Red
+    exit 1
+  }
+  if (-Not(Test-Path -Path $SpriteFile)) {
+    Write-Host "$SpriteFile not found" -ForegroundColor Red
+    exit 1
+  }
 
   Copy-Item -Path "$PathCDTemplate\*" -Destination $PathISOBuildFolder -Recurse -Force -ErrorAction Stop
   Copy-Item -Path $PRGFile -Destination "$PathISOBuildFolder\DEMO.PRG" -Force -ErrorAction Stop
@@ -58,12 +70,12 @@ function Write-SFX {
     [String] $Z80File
   )
 
-  Logger-Step -Message "soundFX"
+  Write-Host "SoundFX" -ForegroundColor Yellow
 
-  if ($PCMFile) { Logger-Info -Message $PCMFile }
-  if ($Z80File) { Logger-Info -Message $Z80File }
+  if ($PCMFile) { Write-Host $PCMFile -ForegroundColor Blue }
+  if ($Z80File) { Write-Host $Z80File -ForegroundColor Blue }
 
-  Logger-Info -Message "Destination folder $PathISOBuildFolder"
+  Write-Host "Destination folder $PathISOBuildFolder" -ForegroundColor Blue
 
   if ($PCMFile) { Copy-Item -Path $PCMFile -Destination "$PathISOBuildFolder\DEMO.PCM" -Force -ErrorAction Stop }
   if ($Z80File) { Copy-Item -Path $Z80File -Destination "$PathISOBuildFolder\DEMO.Z80" -Force -ErrorAction Stop }
@@ -89,34 +101,37 @@ function Write-CUE {
     $ext = [System.IO.Path]::GetExtension($File)
     $path = [System.IO.Path]::GetDirectoryName($File)
 
-    Copy-Item -Path $File -Destination "$($buildConfig.pathBuild)\$path"
+    $destination = "$($Config.project.buildPath)\$($Config.project.name)\$path"
+
+    Copy-Item -Path $File -Destination $destination
 
     if ($ext -eq ".mp3" -or ($ext -eq ".wav" -and $ConfigCDDA.dist.iso.format -eq ".mp3")) {
-      if ((Test-Path -Path "$($buildConfig.pathNeocore)\bin\mpg123-1.31.3-static-x86-64") -eq $false) {
+      if ((Test-Path -Path "$($Config.project.buildPath)\bin\mpg123-1.31.3-static-x86-64") -eq $false) {
         if ($Manifest.manifest.dependencies.mpg123.url) {
           Install-Component `
           -URL $Manifest.manifest.dependencies.mpg123.url `
-          -PathDownload $buildConfig.pathSpool `
+          -PathDownload "$($Config.project.buildPath)\spool" `
           -PathInstall $Manifest.manifest.dependencies.mpg123.path
         } else {
           Install-Component `
-          -URL "$($buildConfig.baseURL)/mpg123-1.31.3-static-x86-64.zip" `
-          -PathDownload $buildConfig.pathSpool `
-          -PathInstall "$($buildConfig.pathNeocore)\bin"
+          -URL "$BaseURL/mpg123-1.31.3-static-x86-64.zip" `
+          -PathDownload "$($Config.project.buildPath)\spool" `
+          -PathInstall "$($Config.project.buildPath)\bin"
         }
       }
     }
 
     if (-Not($Rule -like "*dist*")) {
+      $buildPathProject = "$($Config.project.buildPath)\$($Config.project.name)"
       if ($ext -eq ".wav") {
         Write-Host "copy $File" -ForegroundColor Blue
-        Copy-Item -Path "$($buildConfig.pathBuild)\$path\$baseName$ext" -Destination $path\$baseName$ext
+        Copy-Item -Path "$buildPathProject\$path\$baseName$ext" -Destination $path\$baseName$ext
       }
 
       if ($ext -eq ".mp3") {
         Write-WAV `
-          -mpg123 "$($buildConfig.pathNeocore)\bin\mpg123-1.31.3-static-x86-64\mpg123.exe" `
-          -WAVFile "$($buildConfig.pathBuild)\$path\$baseName.wav" `
+          -mpg123 "$($Config.project.buildPath)\bin\mpg123-1.31.3-static-x86-64\mpg123.exe" `
+          -WAVFile "$buildPathProject\$path\$baseName.wav" `
           -MP3File "$path\$baseName.mp3"
         $File = "$path\$baseName.wav"
       }
@@ -124,63 +139,63 @@ function Write-CUE {
 
     if ($Rule -eq "dist:exe" -and $ext -eq ".mp3") {
       Write-WAV `
-        -mpg123 "$($buildConfig.pathNeocore)\bin\mpg123-1.31.3-static-x86-64\mpg123.exe" `
-        -WAVFile "$($buildConfig.pathBuild)\$path\$baseName.wav" `
+        -mpg123 "$($Config.project.buildPath)\bin\mpg123-1.31.3-static-x86-64\mpg123.exe" `
+        -WAVFile "$buildPathProject\$path\$baseName.wav" `
         -MP3File "$path\$baseName.mp3"
       $File = "$path\$baseName.wav"
     }
 
     if ($Rule -eq "dist:mame" -and $ext -eq ".mp3") {
       Write-WAV `
-        -mpg123 "$($buildConfig.pathNeocore)\bin\mpg123-1.31.3-static-x86-64\mpg123.exe" `
-        -WAVFile "$($buildConfig.pathBuild)\$path\$baseName.wav" `
+        -mpg123 "$($Config.project.buildPath)\bin\mpg123-1.31.3-static-x86-64\mpg123.exe" `
+        -WAVFile "$buildPathProject\$path\$baseName.wav" `
         -MP3File "$path\$baseName.mp3"
       $File = "$path\$baseName.wav"
     }
 
     if ($Rule -eq "dist:exe" -and $ext -eq ".wav") {
       Write-Host "copy $File" -ForegroundColor Blue
-      Copy-Item -Path "$($buildConfig.pathBuild)\$path\$baseName$ext" -Destination $path\$baseName$ext
+      Copy-Item -Path "$buildPathProject\$path\$baseName$ext" -Destination $path\$baseName$ext
     }
 
     if ($Rule -eq "dist:mame" -and $ext -eq ".wav") {
       Write-Host "copy $File" -ForegroundColor Blue
-      Copy-Item -Path "$($buildConfig.pathBuild)\$path\$baseName$ext" -Destination $path\$baseName$ext
+      Copy-Item -Path "$buildPathProject\$path\$baseName$ext" -Destination $path\$baseName$ext
     }
 
     if ($Rule -like "dist:iso") {
       if ($ext -eq ".mp3" -and $ConfigCDDA.dist.iso.format -eq "wav") {
         Write-WAV `
-          -mpg123 "$($buildConfig.pathNeocore)\bin\mpg123-1.31.3-static-x86-64\mpg123.exe" `
-          -WAVFile "$($buildConfig.pathBuild)\$path\$baseName.wav" `
+          -mpg123 "$($Config.project.buildPath)\bin\mpg123-1.31.3-static-x86-64\mpg123.exe" `
+          -WAVFile "$buildPathProject\$path\$baseName.wav" `
           -MP3File "$path\$baseName.mp3"
         $File = "$path\$baseName.wav"
       }
 
       if ($ext -eq ".wav" -and $ConfigCDDA.dist.iso.format -eq "wav") {
         Write-Host "copy $File" -ForegroundColor Blue
-        Copy-Item -Path "$($buildConfig.pathBuild)\$path\$baseName$ext" -Destination $path\$baseName$ext
+        Copy-Item -Path "$buildPathProject\$path\$baseName$ext" -Destination $path\$baseName$ext
       }
 
       if ($ext -eq ".wav" -and $ConfigCDDA.dist.iso.format -eq "mp3") {
-        if ((Test-Path -Path "$($buildConfig.pathNeocore)\bin\ffmpeg.exe") -eq $false) {
+        if ((Test-Path -Path "$($Config.project.buildPath)\bin\ffmpeg.exe") -eq $false) {
           if ($Manifest.manifest.dependencies.ffmpeg.url) {
               Install-Component `
                 -URL $Manifest.manifest.dependencies.ffmpeg.url `
-                -PathDownload $buildConfig.pathSpool `
+                -PathDownload "$($Config.project.buildPath)\spool" `
                 -PathInstall $Manifest.manifest.dependencies.ffmpeg.url
             } else {
               Install-Component `
-                -URL "$($buildConfig.baseURL)/ffmpeg-23-12-18.zip" `
-                -PathDownload $buildConfig.pathSpool `
-                -PathInstall "$($buildConfig.pathNeocore)\bin"
+                -URL "$BaseURL/ffmpeg-23-12-18.zip" `
+                -PathDownload "$($Config.project.buildPath)\spool" `
+                -PathInstall "$($Config.project.buildPath)\bin"
             }
         }
 
         Write-MP3 `
-          -ffmpeg "$($buildConfig.pathNeocore)\bin\ffmpeg.exe" `
+          -ffmpeg "$($Config.project.buildPath)\bin\ffmpeg.exe" `
           -WAVFile "$path\$baseName.wav" `
-          -MP3File "$($buildConfig.pathBuild)\$path\$baseName.mp3"
+          -MP3File "$buildPathProject\$path\$baseName.mp3"
         $File = "$path\$baseName.wav"
       }
     }
@@ -212,7 +227,10 @@ function Write-CUE {
   Remove-Item -Path "$OutputFile.spool" -Force
 
   if ((Test-Path -Path $OutputFile) -eq $true) {
-    Logger-Success -Message "builded CUE $OutputFile"
+    Write-Host "Builded CUE $OutputFile" -ForegroundColor Green
     Write-Host ""
-  } else { Logger-Error -Message "error - $OutputFile was not generated" }
+  } else {
+    Write-Host "$OutputFile was not generated" -ForegroundColor Red
+    exit 1
+  }
 }