+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.7.1
 
+  - fix mak serve:mame
+  - refactor toolchain scripts
   - improve toolchain log
 
 ## 2.7.0